Fix empty host case
If you start with an existing hostname this creates the volume and
source control. If you clear the hostname the connection is of course
drop but volume and source control are both still being called, since
they are not destroyed. Use a failover proxy in this case which does
nothing but keeps the app from crashing. The proper fix would be most
likely to destroy the controls and then re-create them, but as mentioned
before this does not seem to work.